pyproject.toml
[project]
name = "hpm-backend"
version = "0.1.0"
description = "Health Policy Mapper backend API"
authors = [
    {name = "Thomas de Mello Oliva",email = "thomas.mello301@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "google-genai (>=1.26.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [
  { include = "application",    from = "src" },
  { include = "domain",         from = "src" },
  { include = "infrastructure", from = "src" },
  { include = "presentation",   from = "src" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
